pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-agents>=0.2.4",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
